use bindgen to generate libavif-sys
